requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
anthropic>=0.49.0
pypdf>=4.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
